moved LibWrapperService
